# golibrary/static/swagger.json
 {
    "schemes": [
      "https",
      "http"
    ],
    "swagger": "2.0",
    "info": {
      "description": "golibrary ",
      "title": "Swagger Golibrary",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "email": "apiteam@swagger.io"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "version": "1.0.7"
    },
    "host": "golibrary.swagger.io",
    "basePath": "/v2",
    "paths": {
      "/books/book": {
        "post": {
          "description": "create a book from json",
          "produces": [
            "application/json",
            "application/xml"
          ],
          "tags": [
            "book"
          ],
          "summary": "Creates a new book given an author id",
          "operationId": "createBook",
          "parameters": [
            {
              "description": "Created book object",
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Book"
              }
            }
          ],
          "responses": {
            "default": {
              "description": "successful operation"
            }
          }
        }
      },
      "/books/author": {
        "post": {
          "description": "create a author from json",
          "produces": [
            "application/json",
            "application/xml"
          ],
          "tags": [
            "author"
          ],
          "summary": "Add a new author to the database",
          "operationId": "createAuthor",
          "parameters": [
            {
              "description": "Created author object",
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Author"
              }
            }
          ],
          "responses": {
            "default": {
              "description": "successful operation"
            }
          }
        }
      },
      "/books/rent/{bookId}": {
        "patch": {

          "description": "Rent a single book",
          "produces": [
            "application/json",
            "application/xml"
          ],
          "tags": [
            "book"
          ],
          "summary": "Rent a book by ID",
          "operationId": "rentBook",
          "parameters": [
            {
              "type": "integer",
              "format": "int64",
              "description": "ID of book to rent",
              "name": "bookId",
              "in": "path",
              "required": true
            },
            {
              "type": "integer",
              "format": "int64",
              "description": "ID of user",
              "name": "userId",
              "in": "query",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation"
              }
            }
          }
        },
        "/books/return/{bookId}": {
          "patch": {
  
            "description": "Return a book",
            "produces": [
              "application/json",
              "application/xml"
            ],
            "tags": [
              "book"
            ],
            "summary": "Return a book by ID",
            "operationId": "returnBook",
            "parameters": [
              {
                "type": "integer",
                "format": "int64",
                "description": "ID of book to return",
                "name": "bookId",
                "in": "path",
                "required": true
              },
              {
                "type": "integer",
                "format": "int64",
                "description": "ID of user that borrowed the book",
                "name": "userId",
                "in": "query",
                "required": true
              }
            ],
            "responses": {
              "200": {
                "description": "successful operation"
                }
              }
            }
          },
          "/books/listUsers": {
          "get": {
  
            "description": "list users with rented book info",
            "produces": [
              "application/json",
              "application/xml"
            ],
            "tags": [
              "users"
            ],
            "summary": "list users",
            "operationId": "listUsers",
            "parameters": [
              {
                "type": "integer",
                "format": "int64",
                "description": "page",
                "name": "page",
                "in": "query",
                "required": true
              },
              {
                "type": "integer",
                "format": "int64",
                "description": "page_size",
                "name": "page_size",
                "in": "query",
                "required": true
              }
            ],
            "responses": {
              "200": {
                "description": "successful operation"
                }
              }
            }
          },
          "/books/listBooks": {
          "get": {
  
            "description": "list all books with available info",
            "produces": [
              "application/json",
              "application/xml"
            ],
            "tags": [
              "books"
            ],
            "summary": "list books",
            "operationId": "listBooks",
            "parameters": [
              {
                "type": "integer",
                "format": "int64",
                "description": "page",
                "name": "page",
                "in": "query",
                "required": true
              },
              {
                "type": "integer",
                "format": "int64",
                "description": "page_size",
                "name": "page_size",
                "in": "query",
                "required": true
              }
            ],
            "responses": {
              "200": {
                "description": "successful operation"
                }
              }
            }
          },
          "/books/listAuthors": {
          "get": {
  
            "description": "list all authors with their books",
            "produces": [
              "application/json",
              "application/xml"
            ],
            "tags": [
              "author"
            ],
            "summary": "list authors",
            "operationId": "listAuthors",
            "parameters": [
              {
                "type": "integer",
                "format": "int64",
                "description": "page",
                "name": "page",
                "in": "query",
                "required": true
              },
              {
                "type": "integer",
                "format": "int64",
                "description": "page_size",
                "name": "page_size",
                "in": "query",
                "required": true
              }
            ],
            "responses": {
              "200": {
                "description": "successful operation"
                }
              }
            }
          },
...
    },
     "definitions": {
...
      "Book": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "year": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": "string"
          },
          "author": {
           "$ref": "#/definitions/Author"
          }
        },
        "xml": {
          "name": "Book"
        }
      },
      "Author": {
        "type": "object",
        "properties": {
          "books": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Book",
              "xml": {
                "name": "book"
              }
            }
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "example": "doggie"
          },
          "times_ordered": {
            "type": "integer",
            "format": "int32"
          }
        },
        "xml": {
          "name": "Author"
        }
      },
...
    },
    "securityDefinitions": {
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      },
      "golibrary_auth": {
        "type": "oauth2",
        "flow": "implicit",
        "authorizationUrl": "https://golibrary.swagger.io/oauth/authorize"
      }
    },
...
  }